api/pyproject.toml
[project]
name = "mail-triage-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.117.1",
    "google-auth>=2.40.3",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.4",
    "google-genai>=1.39.1",
    "logging>=0.4.9.6",
    "passlib[bcrypt]>=1.7.4",
    "pdf2image>=1.17.0",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.11",
    "pydantic[email]>=2.11.9",
    "pydantic-settings>=2.11.0",
    "pypdf>=5.1.0",
    "pytesseract>=0.3.13",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlmodel>=0.0.31",
    "uvicorn>=0.37.0",
]

[dependency-groups]
dev = [
    "alembic>=1.18.1",
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
]
